Removes unnecessary dep on custom haddock fork.
A previous commit introduced a test that required a version
of haddock located in a github fork. Seeing that haddock has
over a hundred open issues, odds are my commit won't make it
into the main branch any time soon.

To make the test suite pass when all haddock comments are
present, we need to either disable haddock tests by default
or depend on the default version of haddock. The latter is
better, as it helps other people contribute to this project.

This commit removes the dependency on the custom version
of haddock.