Bump version bounds of base dep on elementary demo. Refs #10.
The version bounds on base for the demo are too low (base 4.7).
This commit bumps that to 5. It is unclear whether 5 is the right
choice, but since all other packages have either 5 or none as the
max bound for base, it is a consistent decision. A different
decision could be made globally about that.