Add demo showing how to make an RV sink from an MSF. Refs #11.
This commit adds a demo based on an existing demo, showing how to create
an RV from an MSF. The RV acts as a sink only, so any side effects need
to be made as part of the MSF, but that could be changed.

There is an alternative way of connecting Hails and MSFs, which is
described in the paper "Functional Reactive Programming, Refactored".