Compile with GHC 8.2, 8.4. Disable GHC 8.0. Refs #23.
Currently, GHC 8.0 is failing to compile gtk2hs-buildtools on our
travis setup. We are temporarily disabling the build to go through.

There was a mistake in the branch name that had to be used
for deployment. This is also fixed by this commit.

Modern versions of GHC were not being used to compile Hails, which means we do
not ensure it is up to date for current users. This commit enables 8.2 and
8.4. There is a compilation issue with GHC 8.6, which can be investiagated
separately.