Fix error in minimal WX example. Refs #15.
The minimal example includes an error due to modRW expecting
a modification function with different arguments.

At present I'm finding wx impossible to compile (on my laptop).
I am providing a fix, but it is a "blind" fix: I cannot compile
it, and I cannot try it. Based on the error message it should
be correct but I cannot be sure.