Fix template. Refs #35.
The current template is broken because it imports modules that do not exist, it
refers to functions that have been removed, and it uses operations whose
signature has changed.

This commit fixes the template such that it compiles and adds comments to
indicate what users need to uncomment or define to make a minimal application
work.