Declare a dependency on the tool cabal. Refs #3.
As reported by @peti, the haddock coverage test fails because it needs the tool
cabal, which is not present. Although it is unclear whether that it a correct
way of testing documentation coverage, if the file does need it then the
dependency should be declared.

This change adds a dependency on the tool cabal from the package cabal-install
in the appropriate test suite in the cabal file. That line should be ignored by
versions of Cabal before 2.0, as reported in the cabal documentation:
https://www.haskell.org/cabal/users-guide/developing-packages.html#pkg-field-build-tool-depends