Update to work with Hails 0.6.0. Refs #57.
The example of the image helper fails to build due to depending on an old
version of a package that has been renamed (protected-objects).  Other parts of
Keera Hails have changed as well, other libraries have been renamed, and
modules have been renamed.

This commit updates the example to work with a current version of Hails (0.6.0),
GHC, and Gtk2.