Fix type errors in older versions of hgettext. Refs #98.
The function hgettext:Text.I18N.GetText.bindTextDomain changed
its type signature in version 0.1.5 of the hgettext library,
and hgettext:Text.I18N.GetText.textDomain in version 0.1.6.

This code was written for versions after 0.1.6.

This commit introduces selective compilation clauses for different hgettext
versions and adjusts types accordingly. A minimal constraint on the version of
hgettext >= 0.1.6 is not possible due to it having a low constraint on GHC
(base 4.12) that prevents us from compiling hails with newer versions of GHC.